app/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
et-xmlfile==1.1.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.1.0
openpyxl==3.1.5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.13.0
